requirements.txt
faiss-cpu
streamlit
sentence-transformers
nltk
azure-cognitiveservices-speech
openai
langchain
langchain_community
streamlit-webrtc